Renaming the device level instances file to common name , since it is defined for different DataTypes.
